python/requirements.txt
python-dotenv
ccxt
openai
pandas_ta
selenium
webdriver-manager
Pillow
youtube-transcript-api
streamlit
plotly
schedule